hacks/genai-intro/artifacts/function/requirements.txt
google-cloud-aiplatform
google-cloud-storage
google-cloud-vision
google-cloud-bigquery